+++ cnv/tabs/configuration.py
@@ -16,48 +16,6 @@ class MasterVolume(ctk.CTkFrame):
     This is for playback volume.
     """
 
-
-class SpokenLanguageSelection(ctk.CTkFrame):
-    """
-    The user gets to decide which language they want to hear.  They may also
-    need to decide which translation provider to utilize w/config for that
-    provider.
-    """
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.columnconfigure(0, minsize=125, uniform="baseconfig")
-        self.columnconfigure(1, weight=2, uniform="baseconfig")
-
-        ctk.CTkLabel(
-            self,
-            text="Spoken Language",
-            anchor="e",   
-        ).grid(column=0, row=0, sticky='e')
-
-        current = settings.get_config_key('language', "English")
-        self.language = tk.StringVar(value=current)
-
-        default_engine_combo = ctk.CTkComboBox(
-            self, 
-            variable=self.language,
-            state='readonly',
-            values=list(settings.LANGUAGES.keys())
-        )
-        default_engine_combo.grid(column=1, row=0, sticky='w')
-
-        self.language.trace_add('write', self.change_language)
-       
-    def change_language(self, a, b, c):
-        newvalue = self.language.get()
-        prior = settings.get_config_key('language', "English")
-        settings.set_config_key('language', newvalue)
-        # tempting to just restart
-        if prior and newvalue != prior:
-            log.info(f'Changing language to {newvalue}')
-            # we should immediately translate and localize the UI       
-
-
 class EngineAuthentication(ctk.CTkTabview):
     """
     Collects tabs for configuring authentication for each of the TTS engines.  The 
@@ -348,7 +306,6 @@ class ConfigurationTab(tk.Frame):
         # MasterVolume(self).pack(side="top", fill="x")
         DirectoryChoices(self).pack(side="top", fill="x")
 
-        SpokenLanguageSelection(self).pack(side="top", fill="x")
         ChannelToEngineMap(self).pack(side="top", fill="x")
         SpeakingToggles(self).pack(side="top", fill="x")
         EngineAuthentication(
